Base part done

+++ Hotels.Domain/Queries/LocationQuery.cs
@@ -6,7 +6,7 @@ namespace Hotels.Domain.Queries
 {
     public class LocationQuery : PageQuery
     {
-        public int Distance { get; set; }
+        public float Distance { get; set; }
         public double Longitude { get; set; }
         public double Latitude { get; set; }
     }
